[REF] account_advanced_reconcile: arrange files
(lp:c2c-financial-addons/6.1 rev 24.2.7)
