+++ account_advanced_reconcile/__openerp__.py
@@ -1,7 +1,8 @@
-# -*- coding: utf-8 -*- ##############################################################################
+# -*- coding: utf-8 -*-
+##############################################################################
 #
-#    Author: Nicolas Bessi, Joel Grand-Guillaume
-#    Copyright 2011-2012 Camptocamp SA
+#    Author: Guewen Baconnier
+#    Copyright 2012 Camptocamp SA
 #
 #    This program is free software: you can redistribute it and/or modify
 #    it under the terms of the GNU Affero General Public License as
@@ -28,6 +29,21 @@
  'description': """
 Advanced reconciliation methods for the module account_easy_reconcile.
 
+account_easy_reconcile, which is a dependency, is available in the branch:
+lp:~openerp-community-committers/+junk/account-extra-addons
+This branch is temporary and will soon be merged with the Akretion master
+branch, but the master branch does not already exist. Sorry for the
+inconvenience.
+
+In addition to the features implemented in account_easy_reconcile, which are:
+ - reconciliation facilities for big volume of transactions
+ - setup different profiles of reconciliation by account
+ - each profile can use many methods of reconciliation
+ - this module is also a base to create others reconciliation methods
+    which can plug in the profiles
+ - a profile a reconciliation can be run manually or by a cron
+ - monitoring of reconcilation runs with a few logs
+
 It implements a basis to created advanced reconciliation methods in a few lines
 of code.
 
@@ -37,6 +53,11 @@ Typically, such a method can be:
    or name
  - Reconcile entries if the partner is equal and the ref match with a pattern
 
+And they allows:
+ - Reconciliations with multiple credit / multiple debit lines
+ - Partial reconciliations
+ - Write-off amount as well
+
 A method is already implemented in this module, it matches on entries:
  * Partner
  * Ref on credit move lines should be case insensitive equals to the ref or
@@ -45,9 +66,6 @@ A method is already implemented in this module, it matches on entries:
 The base class to find the reconciliations is built to be as efficient as
 possible.
 
-Reconciliations with multiple credit / debit lines is possible.
-Partial reconciliation are generated.
-You can choose a write-off amount as well.
 
 So basically, if you have an invoice with 3 payments (one per month), the first
 month, it will partial reconcile the debit move line with the first payment, the second
@@ -55,16 +73,18 @@ month, it will partial reconcile the debit move line with 2 first payments,
 the third month, it will make the full reconciliation.
 
 This module is perfectly adapted for E-Commerce business where a big volume of
-move lines and so, reconciliations, is involved and payments often come from
+move lines and so, reconciliations, are involved and payments often come from
 many offices.
+
  """,
  'website': 'http://www.camptocamp.com',
  'init_xml': [],
- 'update_xml': [],
+ 'update_xml': ['easy_reconcile_view.xml'],
  'demo_xml': [],
  'test': [],
  'images': [],
  'installable': True,
  'auto_install': False,
  'license': 'AGPL-3',
+ 'application': True,
 }
