+++ account_advanced_reconcile/easy_reconcile_view.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="utf-8"?>
+<openerp>
+    <data noupdate="0">
+      <record id="view_easy_reconcile_form" model="ir.ui.view">
+          <field name="name">account.easy.reconcile.form</field>
+          <field name="model">account.easy.reconcile</field>
+          <field name="type">form</field>
+          <field name="inherit_id" ref="account_easy_reconcile.account_easy_reconcile_form"/>
+          <field name="arch" type="xml">
+              <page name="information" position="inside">
+                <group colspan="2" col="2">
+                    <separator colspan="4" string="Advanced. Partner and Ref"/>
+                    <label string="Match multiple debit vs multiple credit entries. Allow partial reconcilation.
+The lines should have the partner, the credit entry ref. is matched vs the debit entry ref. or name." colspan="4"/>
+                </group>
+              </page>
+          </field>
+      </record>
+    </data>
+</openerp>
