--- account_advanced_reconcile/wizard/statement_auto_reconcile_view.xml
@@ -1,72 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<openerp>
-    <data>
-
-        <record id="account_automatic_reconcile_view" model="ir.ui.view">
-             <field name="name">Account Automatic Reconcile</field>
-             <field name="model">account.statement.import.automatic.reconcile</field>
-             <field name="type">form</field>
-             <field name="arch" type="xml">
-                <form string="Reconciliation">
-                    <separator string="Reconciliation" colspan="4"/>
-                    <label colspan="4" nolabel="1" string="For an invoice to be considered as paid, the invoice entries must be reconciled with counterparts, usually bank payments. It could also be reconciled with an intermediate account of a payment office (like PayPal, Amazone, ...). 
-                        With this automatic reconciliation functionality, OpenERP makes its own search for entries to reconcile in a series of accounts. It finds entries for each transaction where the id or origin correspond."/>
-                    <newline/>
-                    <group>
-                        <field name="account_ids" colspan="4" domain="[('reconcile','=',True)]"/>
-                        <field name="partner_ids" colspan="4"/>
-                        <field name="invoice_ids" colspan="4" domain="[('type', '=', 'out_invoice'), ('state', '=', 'open')]"/>
-                        <field name="allow_write_off"/>
-                    </group>
-                    <newline/>
-                    <group attrs="{'readonly':[('allow_write_off', '!=', True)]}">
-                        <separator string="Write-Off Move" colspan="4"/>
-                        <field name="writeoff_acc_id" attrs="{ 'required':[('allow_write_off', '=', True)]}"/>
-                        <field name="writeoff_amount_limit" attrs="{ 'required':[('allow_write_off', '=', True)]}"/>
-                        <field name="journal_id" attrs="{ 'required':[('allow_write_off', '=', True)]}"/>
-                    </group>
-                    <separator string ="" colspan="4"/>
-                    <group colspan="2" col="4">
-                         <button special="cancel" string="Cancel" icon="gtk-cancel"/>
-                        <button name="reconcile" string="Reconcile" type="object" icon="terp-stock_effects-object-colorize"/>
-                     </group>
-                </form>
-             </field>
-         </record>
-
-        <record id="action_account_automatic_reconcile" model="ir.actions.act_window">
-             <field name="name">Account Automatic Reconcile</field>
-             <field name="res_model">account.statement.import.automatic.reconcile</field>
-             <field name="type">ir.actions.act_window</field>
-             <field name="view_type">form</field>
-             <field name="view_mode">tree,form</field>
-             <field name="view_id" ref="account_automatic_reconcile_view"/>
-             <field name="target">new</field>
-       </record>
-
-        <menuitem
-            icon="STOCK_EXECUTE"
-            name="Automatic Statement Reconciliation"
-            action="action_account_automatic_reconcile"
-            id="menu_automatic_reconcile"
-            parent="account.menu_finance_periodical_processing"
-            />
-
-        <record id="stat_account_automatic_reconcile_view1" model="ir.ui.view">
-             <field name="name">Automatic reconcile unreconcile</field>
-             <field name="model">account.statement.import.automatic.reconcile</field>
-             <field name="type">form</field>
-             <field name="arch" type="xml">
-                <form string="Statement Reconciliation result">
-                    <field name="reconciled"/>
-                    <newline/>
-                    <group colspan="4" col="6">
-                        <separator colspan="6"/>
-                        <button special="cancel" string="Ok" icon="terp-dialog-close" default_focus="1"/>
-                     </group>
-                </form>
-             </field>
-         </record>
-
-    </data>
-</openerp>
